requirements.txt
opencv-python==4.10.0.84
numpy==1.23.5
vgamepad==0.1.0
neat-python==0.92
mss==9.0.1
dolphin-memory-engine==1.2.0
pygetwindow==0.0.9